moved all view- to view directory updated the references
