+++ src/bic/router.js
@@ -12,7 +12,7 @@ define(function (require) {
   // local modules
 
   var app = require('bic/model-application');
-  var InteractionView = require('bic/view-interaction');
+  var InteractionView = require('bic/view/interaction');
   var uiTools = require('bic/lib/ui-tools');
 
   // this module
