+++ src/bic/view/form/action.js
@@ -9,8 +9,8 @@ define(function (require) {
   // local modules
 
   var app = require('bic/model-application');
-  var FormControls = require('bic/view-form-controls');
-  var FormErrorSummary = require('bic/view-form-error-summary-list-view');
+  var FormControls = require('bic/view/form/controls');
+  var FormErrorSummary = require('bic/view/form/error-summary-list');
 
   // this module
