+++ src/bic/view/form/list.js
