+++ src/bic/view/star.js
