+++ tests/unit/view-star.js
@@ -9,7 +9,7 @@ define(['Squire', 'chai'], function (Squire, chai) {
     before(function (done) {
       injector = new Squire();
 
-      injector.require(['bic/view-star'], function (rView) {
+      injector.require(['bic/view/star'], function (rView) {
         View = rView;
         done();
       });
